prototype/requirements.txt
coreapi>=2.3.3
coreschema>=0.0.4
Django>=2.0.7
django-allauth>=0.36.0
django-cors-headers>=2.4.0
django-rest-auth>=0.9.3
djangorestframework>=3.8.2
djangorestframework-jwt>=1.11.0
oauthlib>=2.1.0
protobuf>=3.6.1
pylint>=1.8.4
python-dateutil>=2.7.3
psycopg2-binary>=2.7.5
requests>=2.19.1
requests-oauthlib>=1.0.0
uritemplate>=3.0.0
urllib3>=1.22
websocket>=0.2.1
